fix #259: Handlebars.Utils.isEmpty false positive for small big decimal

+++ handlebars/src/main/java/com/github/jknack/handlebars/Handlebars.java
@@ -193,7 +193,7 @@ public class Handlebars implements HelperRegistry {
         return Array.getLength(value) == 0;
       }
       if (value instanceof Number) {
-        return ((Number) value).intValue() == 0;
+        return ((Number) value).doubleValue() == 0;
       }
       return false;
     }
